Support binding_deactivate in CacheBackedPluginApi
CacheBackedPluginApi enables the neutron server to push resources
updates to L2 agents. The agents retrieve the resources updates locally
from the cache implemented by it. CacheBackedPluginApi also emulates
server notifications to the agents, such as port_delete or port_update,
based on the updated data received by the cache. This commit adds code
to CacheBackedPluginApi to implement a binding_deactivate notification
from the server to the agents

Change-Id: I023ccbd405bc41379007d87a9c1051970aa8d603
Partial-Bug: #1580880