Improve DbListCommand operation from O(n^2) to O(n)
Right now the DbListCommand retrieves the uuid's of all the elements
passed in as argument. This is an O(n^2) operation so when the number
of elements in a grows it's likely that we get Timeout Exceptions.

Instead of doing this, whenever possible, we'll retrieve all the
elements (from the in-memory replica) and only fetch those who were
passed as arguments avoiding the O(n^2) operation.

Closes-Bug: #1769897
Change-Id: I04bfe243c21017c52834167e6e491f64c516c29d
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>