doc: update doc to refer new trusted tag for SR-IOV
It allows Virtual Functions to become “trusted” by the Physical
Function and perform some privileged operations, such as enabling VF
promiscuous mode and changing VF MAC address within the guest. The
inability to modify MAC addresses in the guest prevents the users from
being able to easily set up two VFs in a fail-over bond in a
guest. This spec aims to suggest a way for users to boot instances
with trusted VFs.

https://specs.openstack.org/openstack/nova-specs/specs/rocky/approved/sriov-trusted-vfs.html

Change-Id: I58cbfd9698a6c7832abc52e0be70eab72f66e11a
Depends-On: https://review.openstack.org/#/c/458820/
Signed-off-by: Sahid Orentino Ferdjaoui <sahid.ferdjaoui@redhat.com>