Mention neutron-rootwrap-daemon in root_helper_daemon option help
An operators mailing list thread [1] discussing scale issues
where vif plugging times out eventually came to the conclusion
that several operators run rootwrap in the neutron agent in daemon
mode.

This ability is not clear from the configuration option help, however,
so this change adds more details on running in this mode to achieve
scale. Another nice discussion about this can be found here [2].

There is a corresponding nova patch [3] to mention this option
if operators are hitting vif plugging timeouts in nova.

This doesn't change any defaults, it just improves the documentation.

The help text formatting is also updated for easier readability
and maintainability in the source.

[1] http://lists.openstack.org/pipermail/openstack-operators/2018-May/015364.html
[2] https://cloudblog.switch.ch/2017/08/28/starting-1000-instances-on-switchengines/
[3] https://review.openstack.org/571577/

Change-Id: I8b8a0cfea409c3469ad0ee223c87e0380b3afbad
Closes-Bug: #1774217