Fix unscheduling fail when concurrent port delete
When a network is removed from a dhcp agent, in some scenarios if the
agent releases its port concurrently, there is chance that the
unscheduling will fail due to that the target port is not found.

Catch the PortNotFound exception as an expected error under this type of
concurrent circumstance and logs it to move forward.

Closes-Bug: #1775496
Change-Id: Ib51b364f6ced0de7685c8ee07c1d292308d919f5
Signed-off-by: Kailun Qin <kailun.qin@intel.com>