Fix DVR scheduling checks wrong profile for host
When DVR scheduling in live-migration, the current implementation in DVR
serviceable ports checking on host performs a "contains" operation of
sub-string match which checks the wrong port binding profile for host
(i.e. compute-1 will also match compute-10).

Add quotes to force an exact match of the host name in the port binding
profile dictionary to address this issue.

Closes-Bug: #1776255
Change-Id: I0d2bd9b9ff0aa58a7cce1b8da2a5f21ac6b38c57
Signed-off-by: Kailun Qin <kailun.qin@intel.com>