Mark test_ha_router_restart_agents_no_packet_lost as unstable
Fullstack test
test_l3_agent.TestHAL3Agent.test_ha_router_restart_agents_no_packet_lost
is marked as unstable now becuase it is failing quite often recently.
We need to figure out what is the reason of this issue but
to not block gates with many failures, let's mark it as unstable
for now.

Change-Id: I21e590a24390345dfe451b035fd973928445e987
Related-Bug: #1776459