Use rally 0.12.1 release for stable/ocata branch
Rally does not have any alignment of own tags to OpenStack releases.
In theory it should be possible to launch the latest Rally against
old clouds.

To follow the common processes and policy, Rally synchronized own
dependencies (python-*client's and oslo.* libs) with g-r. It made a
conflict between installed versions by stable/pike@devstack and rally
itself.
With Rally master branch it became an error (Rally's team is
working on moving in-tree OpenStack plugins to the separate repository
openstack/rally-openstack).
Rally 0.12.1 is the latest release which doesn't contain a breaking
change.

Ideally, to avoid such issues in future, Rally devstack plugin and
zuul job should install Rally in a virtual environment, but such
change requires more time for investigation and implementation.

So this patch is a quick workaround to avoid version conflict
of python dependencies.

Change-Id: I73ef5c978e66a89bb70be28468e0d1f40a71d94f
Closes-Bug: #1777506