[Docs] Add note about Floating IP QoS
As it might be confusing for users that ID of QoS policy
attached to FIP will be also visible in port's attributes
after associate FIP with port this patch adds small note
about that in QoS documentation.

This patch clarify also rules from which QoS policy will
be instelled by L3 agent in case when different QoS polies
are associated with a port and a floating IP.

Additionally this commit fixes also one typo found in
qos configs document.

Change-Id: I13e49a12be61407fa0f2898bef7deda64a04a1f5
Closes-Bug: 1781892