Class to represent Placement state and sync
Represent the desired state of the Placement DB and provide a method to
sync this state to Placement. This is unused as is. A later patch in
the same series is going to use it. But it is broken out into its own
file and patch, so the code is organized better and easier to review.

Change-Id: I3728117f90b7fad2423c2a4276db214ec86c9d8e
Depends-On: https://review.openstack.org/570847
Depends-On: https://review.openstack.org/577220
Depends-On: https://review.openstack.org/616194
Partial-Bug: #1578989
See-Also: https://review.openstack.org/502306 (nova spec)
See-Also: https://review.openstack.org/508149 (neutron spec)