Use nested virt in scenario jobs
virt_type option in nova-compute will be now set to "kvm" instead
of "qemu" if test job will be running on node which supports
nested virtualization.
In case of nodes where it's not supported, devstack will
automatically switch it to "qemu" again. It's in [1].

This should improve time of booting vms so tests should be finished
faster and there should be less errors with ssh to instance timeouts.

[1]
http://git.openstack.org/cgit/openstack-dev/devstack/tree/lib/nova#n289

Change-Id: Ia361f33f0b103ed44ef372c1464b3c0473a19291