Wait until all router ports are DOWN before migration
In router migration tests, before migration is started, router
is set to admin_state_up=False. This cause that status of all
router ports should be set to DOWN.
This patch adds check (and wait) that all ports are really set
to DOWN state before migration of router is started.

Change-Id: I93db6f67a74f753eaad0900e8045d4676dd1337c
Related-Bug: #1785582