Correctly call is_distributed_router()
A recent change to _notify_l3_agent_port_update() was
making a call to is_distributed_router() via the l3plugin,
but since that method lives in l3_dvr_db.py it was
failing.  This was not caught by the unit test since it
was mocking the return value.

Had to modify things to get the router first since the
object is required for the check.

Change-Id: I2289f882b692b375d1b40d7a50d206d72da66155
Closes-bug: #1785848