shim model query hooks to use neutron-lib
This patch shims the model query hook functionality to use neutron-lib.
A shim is used as the query hooks are cached in a shared dict and by
shimming them to use neutron-lib we can let consumers move over to
neutron-lib's version at will. Once all current consumers are moved
over to lib, we can remove the shim and consume from lib directly.

The DBQueryHooksFixture is also used in this patch to replace the
existing cleanup function in the unit tests.

Change-Id: I18a4e836ce89b607b5452527e5f8dd172cad6b4f