Add argument to get ports by SG to allow excluding owners
When code needs to load a list of all ports that are not router
interfaces, they tend to fetch the port information into memory and then
iterate over the ports, inspecting each port's device_owner.  For large
numbers of ports this is presumably inefficient.  This came up in a
discussion of just such a part in the code [1].

A more efficient solution is to perform this filtering in SQL.  This
change adds an argument to an existing method in the Port object to do
that.

[1] https://review.openstack.org/580721

Change-Id: I742d8a79175209e42f8bd6319956e960c73201ee