Parse dhcp leases file in a more robust way
It turns out that in environments with a big number of VMs, sometimes
the neutron dhcp agent fails to read the dhcp lease file because some
lines with the ipv4/ipv6 entries don't have enough fields and causes the
dhcp agent to fail.

When this happens the agent calls sync_state to
fully resync the agent state, that causes a serious performance problems
in scale environments.

We need to be more robust reading the file to handle these exceptions.

Co-authored-by: stephen-ma
Partial-Bug: #1788556

Change-Id: Ia681a5e929df5bf8c97ae9445876c306c34061b5