Use older keepalived version in functional tests
It looks that current keepalived version which is available in
Ubuntu Xenial repositories (1:1.2.24-1ubuntu0.16.04.1) is broken
and cause failure of some functional tests in Neutron.
Details are in [1].
Older version works fine so as temporary solution we can use
this version in functional tests.

This issue don't happens on master and stable/rocky branch, as there
newer cloud-archive repo is used and it has newer version of keepalived
which works fine.

[1] https://bugs.launchpad.net/ubuntu/+source/keepalived/+bug/1789045

Change-Id: I418a967cd503991736e72134d4a105b6e97021e8
Closes-Bug: #1788185