Fix devstack/lib/ovs to run with Fedora 28
The patch is fixing two problems found when stacking DevStack on a
Fedora 28 host OS.

Problem 1: Account to the different patch versions between the
kernel-devel and kernel-headers packages.

Problem 2: Install the elfutils-libelf-devel package which is needed to
compile OVS.

For more a detailed information about each problem, check the bug linked
in this patch.

Closes-Bug: #1790143
Change-Id: Idfdee28124ff19272abcaaa3adade0435e3e474a