Exclude fallback tunnel devices from netns cleanup
When a namespace gets created, if the fb_tunnels_only_for_init_net
sysctl configuration is set to 0, the fallback tunnels will be
automatically created if the corresponding tunnel is present
in the initial network namespace. Source [0].

This is being observed in some systems where namespaces are not
getting cleaned up due to the presence of such devices as
ip_lib.namespace_is_empty() is returning False.

This patch is adding such devices as per the kernel documentation
to the list of excluded devices by default.

[0] https://www.kernel.org/doc/Documentation/sysctl/net.txt

Closes-Bug: #1797084
Change-Id: I94415a0da5367e2d98d792a5eb4ba3919b838326
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>
(cherry picked from commit 8e60531d11401c2c509a1255b49f89bc2bf81a74)